0.6.7 Website - Refactor Final
